reduce the number of jq spawns by make
When a make variable is not set, 'VAR?= content' sets it as a
"recursively expanded" variable, meaning that the expansion of any
other variables or functions referenced in its content is delayed until
this variable is expanded.  When such a variable is exported, the
expansion of its content also occurs just before each subprocess
creation for which a corresponding environment variable has to be set.

In this Makefile this resulted in a large number of 'jq' processes
being spawned, which is costly on (some?) Windows systems.

Avoid this by transforming the variables into simply expanded variables
after the conditional assignments (and export them at this point).